+++ api/tests/apis/v1/course/test_course_command_permission.py
