+++ api/tests/apis/v1/course/test_course_commands_api.py
