+++ api/tests/apis/v1/course/test_course_query_permission.py
